move spec to backend-base
