+++ backend-base/spec/helpers/retina_image_helper_spec.rb
