+++ backend-base/spec/requests/content_security_policy_header_spec.rb
