+++ backend-base/spec/requests/http_basic_auth_middleware_spec.rb
