+++ backend-base/template.rb
@@ -1,6 +1,7 @@
 source_paths.unshift(File.dirname(__FILE__))
 
+apply "spec/template.rb"
 # The accessibility template brings in the lighthouse and
 # lighthouse matcher parts we need to run performance specs
-apply "backend-base/accessibility/template.rb"
-apply "backend-base/performance/template.rb"
+apply "accessibility/template.rb"
+apply "performance/template.rb"
